% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parTemperBest.R
\name{parTemperBest}
\alias{parTemperBest}
\title{Get best value from call to parTemper}
\usage{
parTemperBest(input, chainsOnly = F)
}
\arguments{
\item{input}{The output of parTemper or a list of chains, chainList}

\item{chainsOnly}{An optional boolean indicating whether the input is a list of chains only [default FALSE]}
}
\value{
The parameter X with the lowest value of the cost function
}
\description{
Get the best value (the X with the lowest associated value of the cost function) from the output of parTemper. The optional input chainsOnly [default FALSE] allows a set of chains to be passed to parTemperBest rather than the output of parTemper and is used internally by parTemper.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
