% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saTemper.R
\name{saTemper}
\alias{saTemper}
\title{Parallel tempering for function minimization}
\usage{
saTemper(costFunc, init, tempVect = NA, ..., control = list())
}
\description{
Parallel tempering for function minimization
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
