% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saTemperBest.R
\name{saTemperBest}
\alias{saTemperBest}
\title{Get best value from call to saTemper}
\usage{
saTemperBest(input, chainsOnly = F)
}
\description{
Get best value from call to saTemper
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
