% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saTemper (mhpricemsi's conflicted copy
%   2019-05-15).R, R/saTemper.R
\name{saTemper}
\alias{saTemper}
\title{Parallel tempering for function minimization}
\usage{
saTemper(costFunc, init, ..., control = list())

saTemper(costFunc, init, ..., control = list())
}
\description{
Parallel tempering for function minimization

Parallel tempering for function minimization
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>

Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
