% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saTemperBest.R
\name{saTemperBest}
\alias{saTemperBest}
\title{Get best value from call to saTemper}
\usage{
saTemperBest(input, chainsOnly = F)
}
\arguments{
\item{input}{The output of saTemper or a list of chains, chainList}

\item{chainsOnly}{An optional boolean indicating whether the input is a list of chains only [default FALSE]}
}
\value{
The parameter X with the lowest value of the cost function
}
\description{
Get the best value (the X with the lowest associated value of the cost function) from the output of saTemper. The optional input chainsOnly [default FALSE] allows a set of chains to be passed to saTemperBest rather than the output of saTemper and is used internally by saTemper.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
